Moving necessary common preferences for review
