+++ Client/Extensions/SceneExtensions.swift
@@ -26,41 +26,3 @@ extension UIWindowScene {
     return browserViewControllers.first
   }
 }
-
-extension UIView {
-  /// Returns the `Scene` that this view belongs to.
-  /// If the view does not belong to a scene, it returns the scene of its parent
-  /// Otherwise returns nil if no scene is associated with this view.
-  var currentScene: UIWindowScene? {
-    if let scene = window?.windowScene {
-      return scene
-    }
-    
-    if let scene = superview?.currentScene {
-      return scene
-    }
-    
-    return nil
-  }
-}
-
-extension UIViewController {
-  /// Returns the `Scene` that this controller belongs to.
-  /// If the controller does not belong to a scene, it returns the scene of its presenter or parent.
-  /// Otherwise returns nil if no scene is associated with this controller.
-  var currentScene: UIWindowScene? {
-    if let scene = view.window?.windowScene {
-      return scene
-    }
-    
-    if let scene = parent?.currentScene {
-      return scene
-    }
-    
-    if let scene = presentingViewController?.currentScene {
-      return scene
-    }
-    
-    return nil
-  }
-}
